pyproject.toml
[tool.poetry]
name = "swiss-army-man"
version = "0.1.19"
description = ""
authors = ["Brett Shollenberger <brettshollenberger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.4.2"
matplotlib = "^3.8.4"
pandas = "^2.2.2"
numpy = "^1.26.4"
datetime = "^5.5"
redis = "^5.0.3"
tqdm = "^4.66.2"
twine = "^5.0.0"
pyyaml = "^6.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
